requirements.txt
flask==2.3.3
pillow==12.1.0
gunicorn==21.2.0
pyopenssl==24.0.0
pyzbar==0.1.9